Revert things with Context dividable commands
