--- tg-bot/src/main/java/polis/commands/CommandRegistry.java
@@ -1,7 +1,0 @@
-package polis.commands;
-
-import org.telegram.telegrambots.extensions.bots.commandbot.commands.IBotCommand;
-
-public interface CommandRegistry<C extends IBotCommand> {
-    C getRegisteredCommand(String commandIdentifier);
-}
