+++ tg-bot/src/main/java/polis/commands/impl/AddOkAccount.java
@@ -1,4 +1,4 @@
-package polis.commands.contextfull;
+package polis.commands.impl;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -9,27 +9,27 @@ import org.telegram.telegrambots.meta.bots.AbsSender;
 import polis.commands.Command;
 import polis.keyboards.InlineKeyboard;
 import polis.keyboards.ReplyKeyboard;
+import polis.ok.api.OkAuthorizator;
 import polis.util.State;
-import polis.vk.api.VkAuthorizator;
 
 import java.net.URISyntaxException;
 
 @Component
-public class AddVkAccount extends Command {
-    private static final String VK_AUTH_ANSWER_MSG = """
-                    Для авторизации в социальной сети ВКонтакте перейдите по ссылке:
+public class AddOkAccount extends Command {
+    private static final String OK_AUTH_ANSWER_MSG = """
+                    Для авторизации в социальной сети Одноклассники перейдите по ссылке:
                     %s
-                    После авторизации скопируйте всю ссылку из адресной строки целиком и отправьте ее в этот диалог.""";
-    private static final Logger LOGGER = LoggerFactory.getLogger(AddVkAccount.class);
+                    После авторизации скопируйте код авторизации из адресной строки и отправьте его в этот диалог.""";
+    private static final Logger LOGGER = LoggerFactory.getLogger(AddOkAccount.class);
 
-    public AddVkAccount(InlineKeyboard inlineKeyboard, ReplyKeyboard replyKeyboard) {
-        super(State.AddVkAccount.getIdentifier(), State.AddVkAccount.getDescription(), inlineKeyboard, replyKeyboard);
+    public AddOkAccount(InlineKeyboard inlineKeyboard, ReplyKeyboard replyKeyboard) {
+        super(State.AddOkAccount.getIdentifier(), State.AddOkAccount.getDescription(), inlineKeyboard, replyKeyboard);
     }
 
     @Override
     public void execute(AbsSender absSender, User user, Chat chat, String[] arguments) {
         try {
-            String messageText = String.format(VK_AUTH_ANSWER_MSG, VkAuthorizator.formAuthorizationUrl());
+            String messageText = String.format(OK_AUTH_ANSWER_MSG, OkAuthorizator.formAuthorizationUrl());
             sendAnswerWithOnlyBackButton(
                     absSender,
                     chat.getId(),
