+++ tg-bot/src/main/java/polis/commands/impl/AddVkGroup.java
@@ -1,4 +1,4 @@
-package polis.commands.contextfull;
+package polis.commands.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
