+++ tg-bot/src/main/java/polis/commands/impl/StartCommand.java
@@ -1,11 +1,11 @@
-package polis.commands.contextless;
+package polis.commands.impl;
 
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.objects.Chat;
 import org.telegram.telegrambots.meta.api.objects.User;
 import org.telegram.telegrambots.meta.bots.AbsSender;
 import polis.commands.Command;
-import polis.commands.ContextLessCommand;
+import polis.commands.DescribableCommand;
 import polis.keyboards.InlineKeyboard;
 import polis.keyboards.ReplyKeyboard;
 import polis.util.IState;
@@ -14,7 +14,7 @@ import polis.util.State;
 import java.util.List;
 
 @Component
-public class StartCommand extends Command implements ContextLessCommand {
+public class StartCommand extends Command implements DescribableCommand {
     private static final String NEXT_COMMANDS_DESCRIPTION = """
             Введите /%s и добавьте новый Телеграмм-канал, из которого хотите публиковать посты в другие социальные сети.
             Справка по боту доступна по команде /%s.
